⚙️ Move unit tests to different folder
