+++ server/test/unit/controllers/artistController.test.js
@@ -1,12 +1,12 @@
-const app = require("../../src/app");
-const User = require("../../src/models/userModel");
-const RefreshToken = require("../../src/models/refreshTokenModel");
+const app = require("../../../src/app");
+const User = require("../../../src/models/userModel");
+const RefreshToken = require("../../../src/models/refreshTokenModel");
 const request = require("supertest");
 const {
   connectToDatabase,
   cleanupDatabaseAndDisconnect,
-} = require("../helpers/databaseHelpers");
-const createUsersAndLogin = require("../helpers/createUsersAndLogin");
+} = require("../../helpers/databaseHelpers");
+const createUsersAndLogin = require("../../helpers/createUsersAndLogin");
 
 let server;
 beforeAll(async () => {
