+++ server/test/unit/controllers/src/cat.jpg
