+++ server/test/unit/controllers/userController.test.js
@@ -1,16 +1,16 @@
 const mongoose = require("mongoose");
-const app = require("../../src/app");
+const app = require("../../../src/app");
 const request = require("supertest");
-const User = require("../../src/models/userModel");
-const RefreshToken = require("../../src/models/refreshTokenModel");
+const User = require("../../../src/models/userModel");
+const RefreshToken = require("../../../src/models/refreshTokenModel");
 const fs = require("node:fs");
-const { resolve } = require("node:path");
-const signupAndLoginUser = require("../helpers/signupAndLoginUser");
+const {resolve} = require("node:path");
+const signupAndLoginUser = require("../../helpers/signupAndLoginUser");
 const {
   connectToDatabase,
   cleanupDatabaseAndDisconnect,
-} = require("../helpers/databaseHelpers");
-const createUsersAndLogin = require("../helpers/createUsersAndLogin");
+} = require("../../helpers/databaseHelpers");
+const createUsersAndLogin = require("../../helpers/createUsersAndLogin");
 
 let server;
 beforeAll(async () => {
@@ -94,14 +94,14 @@ describe("userController", () => {
     let accessToken, img;
 
     beforeAll(async () => {
-      ({ accessToken, img } = await signupAndLoginUser());
+      ({accessToken, img} = await signupAndLoginUser());
     });
 
     it("should update user name", async () => {
       const res = await request(app)
         .patch("/api/users/updateMe")
         .set("Cookie", [`accessToken=${accessToken}`])
-        .send({ name: "Jane Doe" });
+        .send({name: "Jane Doe"});
 
       expect(res.status).toBe(200);
       expect(res.body.status).toBe("success");
